app/requirements.txt
fastapi
uvicorn
pydantic-settings
python-json-logger
sqlalchemy
asyncpg
aioredis
celery
pytest
pytest-asyncio
httpx
aiosqlite
pytest-cov
alembic
prometheus-fastapi-instrumentator
passlib[bcrypt]
PyJWT
python-multipart
argon2-cffi